Update structure for tests
separation of concerns
